Extract HandleBars helpers to separate file
