+++ package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "babel-plugin-handlebars-inline-precompile": "^2.1.1",
+    "babel-runtime": "^6.22.0",
     "html-entities": "^1.2.1"
   }
 }
